feat: 네이버 지도 컴포넌트로 변경 - 재사용 가능하도록 변수 설정
